Add files via upload
The second and final project for the class. The goal of this project is to build a prediction model that can predict ambient air pollution concentrations across the continental United States.
We compared 4 models: Linear Regression, K-Nearest Neighbors, Decision Trees and Random Forest to find what model had the best predictions based on their RMSE score.